+++ NEWS.md
@@ -1,3 +1,11 @@
+# BFpack 1.3.0
+
+* Date: 2024-06-11
+* Package update for implementation step in JASP.
+* cor_test now support ordinal measurement levels.
+* bug fixes in Fortran subroutine for cor_test
+
+
 # BFpack 1.2.4
 
 * Date: 2023-11-16